Transient TLS client certificates work.
However, there's no management interface to review and delete client
certificates yet, though they do not persist beyond a restart of Gneto.

Some code was reorganized and renamed, with a number of new existing
functions moving to the new certificates.go file.

Also, a minor bug fix to the regular expressions for Gemini headings.
Headings without a space after the '#' were not interpreted correctly.

Also, the --verbose and --debug CLI flags were replaced with the new
--loglevel, which accepts an integer argument. `--loglevel 3` is
equivalent to the old --debug.